updated the Kafka+SQL storage to support multi-tenancy
