+++ storage/kafkasql/src/main/java/io/apicurio/registry/storage/impl/kafkasql/KafkaSqlRegistryStorage.java
@@ -63,6 +63,7 @@ import io.apicurio.registry.content.extract.ContentExtractor;
 import io.apicurio.registry.logging.Logged;
 import io.apicurio.registry.metrics.PersistenceExceptionLivenessApply;
 import io.apicurio.registry.metrics.PersistenceTimeoutReadinessApply;
+import io.apicurio.registry.mt.TenantContext;
 import io.apicurio.registry.rest.beans.ArtifactSearchResults;
 import io.apicurio.registry.rest.beans.EditableMetaData;
 import io.apicurio.registry.rest.beans.SearchOver;
@@ -130,6 +131,9 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
     KafkaSqlStore sqlStore;
 
     @Inject
+    TenantContext tenantContext;
+
+    @Inject
     @ConfigProperty(name = "registry.kafkasql.bootstrap.servers")
     String bootstrapServers;
 
@@ -144,7 +148,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
     @Inject
     @ConfigProperty(name = "registry.kafkasql.consumer.poll.timeout", defaultValue = "1000")
     Integer pollTimeout;
-    
+
     @Inject
     ArtifactTypeUtilProviderFactory factory;
 
@@ -276,7 +280,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
         String contentHash = DigestUtils.sha256Hex(contentBytes);
         
         if (!sqlStore.isContentExists(contentHash)) {
-            CompletableFuture<UUID> future = submitter.submitContent(artifactId, contentHash, artifactType, content);
+            CompletableFuture<UUID> future = submitter.submitContent(tenantContext.tenantId(), artifactId, contentHash, artifactType, content);
             UUID uuid = ConcurrentUtil.get(future);
             coordinator.waitForResponse(uuid);
         }
@@ -314,7 +318,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
         }
 
         return submitter
-                .submitArtifact(artifactId, ActionType.Create, artifactType, contentHash, createdBy, createdOn, metaData)
+                .submitArtifact(tenantContext.tenantId(), artifactId, ActionType.Create, artifactType, contentHash, createdBy, createdOn, metaData)
                 .thenCompose(reqId -> (CompletionStage<ArtifactMetaDataDto>) coordinator.waitForResponse(reqId));
     }
 
@@ -327,7 +331,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
             throw new ArtifactNotFoundException(artifactId);
         }
 
-        UUID reqId = ConcurrentUtil.get(submitter.submitArtifact(artifactId, ActionType.Delete));
+        UUID reqId = ConcurrentUtil.get(submitter.submitArtifact(tenantContext.tenantId(), artifactId, ActionType.Delete));
         SortedSet<Long> versionIds = (SortedSet<Long>) coordinator.waitForResponse(reqId);
 
         return versionIds;
@@ -367,7 +371,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
         }
         
         return submitter
-                .submitArtifact(artifactId, ActionType.Update, artifactType, contentHash, createdBy, createdOn, metaData)
+                .submitArtifact(tenantContext.tenantId(), artifactId, ActionType.Update, artifactType, contentHash, createdBy, createdOn, metaData)
                 .thenCompose(reqId -> (CompletionStage<ArtifactMetaDataDto>) coordinator.waitForResponse(reqId));
     }
 
@@ -420,7 +424,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
         // Note: the next line will throw ArtifactNotFoundException if the artifact does not exist, so there is no need for an extra check.
         ArtifactMetaDataDto metaDataDto = sqlStore.getArtifactMetaData(artifactId);
         
-        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactVersion(artifactId, metaDataDto.getVersion(), ActionType.Update, metaDataDto.getState(), metaData));
+        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactVersion(tenantContext.tenantId(), artifactId, metaDataDto.getVersion(), ActionType.Update, metaDataDto.getState(), metaData));
         coordinator.waitForResponse(reqId);
     }
 
@@ -442,7 +446,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
         }
 
         return submitter
-                .submitArtifactRule(artifactId, rule, ActionType.Create, config)
+                .submitArtifactRule(tenantContext.tenantId(), artifactId, rule, ActionType.Create, config)
                 .thenCompose(reqId -> {
                     CompletionStage<Void> rval = (CompletionStage<Void>) coordinator.waitForResponse(reqId);
                     log.debug("===============> Artifact rule (async) completed.  Rval: {}", rval);
@@ -461,9 +465,9 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
 
         // TODO find a better way to implement deleting all rules?  This isn't very scalable if we add more rules.
         
-        submitter.submitArtifactRule(artifactId, RuleType.COMPATIBILITY, ActionType.Delete);
+        submitter.submitArtifactRule(tenantContext.tenantId(), artifactId, RuleType.COMPATIBILITY, ActionType.Delete);
         
-        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactRule(artifactId, RuleType.VALIDITY, ActionType.Delete));
+        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactRule(tenantContext.tenantId(), artifactId, RuleType.VALIDITY, ActionType.Delete));
         try {
             coordinator.waitForResponse(reqId);
         } catch (RuleNotFoundException e) {
@@ -488,7 +492,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
             throw new RuleNotFoundException(rule);
         }
 
-        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactRule(artifactId, rule, ActionType.Update, config));
+        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactRule(tenantContext.tenantId(), artifactId, rule, ActionType.Update, config));
         coordinator.waitForResponse(reqId);
     }
 
@@ -501,7 +505,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
             throw new RuleNotFoundException(rule);
         }
 
-        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactRule(artifactId, rule, ActionType.Delete));
+        UUID reqId = ConcurrentUtil.get(submitter.submitArtifactRule(tenantContext.tenantId(), artifactId, rule, ActionType.Delete));
         coordinator.waitForResponse(reqId);
     }
 
@@ -551,7 +555,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
     @Override
     public void deleteArtifactVersion(String artifactId, long version) throws ArtifactNotFoundException, VersionNotFoundException, RegistryStorageException {
         handleVersion(artifactId, version, null, value -> {
-            UUID reqId = ConcurrentUtil.get(submitter.submitVersion(artifactId, (int) version, ActionType.Delete));
+            UUID reqId = ConcurrentUtil.get(submitter.submitVersion(tenantContext.tenantId(), artifactId, (int) version, ActionType.Delete));
             coordinator.waitForResponse(reqId);
             return null;
         });
@@ -563,7 +567,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
     @Override
     public void updateArtifactVersionMetaData(String artifactId, long version, EditableArtifactMetaDataDto metaData) throws ArtifactNotFoundException, VersionNotFoundException, RegistryStorageException {
         handleVersion(artifactId, version, ArtifactStateExt.ACTIVE_STATES, value -> {
-            UUID reqId = ConcurrentUtil.get(submitter.submitArtifactVersion(artifactId, (int) version, ActionType.Update, value.getState(), metaData));
+            UUID reqId = ConcurrentUtil.get(submitter.submitArtifactVersion(tenantContext.tenantId(), artifactId, (int) version, ActionType.Update, value.getState(), metaData));
             return coordinator.waitForResponse(reqId);
         });
     }
@@ -574,7 +578,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
     @Override
     public void deleteArtifactVersionMetaData(String artifactId, long version) throws ArtifactNotFoundException, VersionNotFoundException, RegistryStorageException {
         handleVersion(artifactId, version, null, value -> {
-            UUID reqId = ConcurrentUtil.get(submitter.submitVersion(artifactId, (int) version, ActionType.Clear));
+            UUID reqId = ConcurrentUtil.get(submitter.submitVersion(tenantContext.tenantId(), artifactId, (int) version, ActionType.Clear));
             return coordinator.waitForResponse(reqId);
         });
     }
@@ -613,7 +617,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
      */
     @Override
     public void createGlobalRule(RuleType rule, RuleConfigurationDto config) throws RuleAlreadyExistsException, RegistryStorageException {
-        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(rule, ActionType.Create, config));
+        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(tenantContext.tenantId(), rule, ActionType.Create, config));
         coordinator.waitForResponse(reqId);
     }
 
@@ -624,9 +628,9 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
     public void deleteGlobalRules() throws RegistryStorageException {
         // TODO find a better way to implement deleting all global rules?  This isn't very scalable if we add more rules.
         
-        submitter.submitGlobalRule(RuleType.COMPATIBILITY, ActionType.Delete);
+        submitter.submitGlobalRule(tenantContext.tenantId(), RuleType.COMPATIBILITY, ActionType.Delete);
         
-        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(RuleType.VALIDITY, ActionType.Delete));
+        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(tenantContext.tenantId(), RuleType.VALIDITY, ActionType.Delete));
         try {
             coordinator.waitForResponse(reqId);
         } catch (RuleNotFoundException e) {
@@ -651,7 +655,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
             throw new RuleNotFoundException(rule);
         }
         
-        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(rule, ActionType.Update, config));
+        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(tenantContext.tenantId(), rule, ActionType.Update, config));
         coordinator.waitForResponse(reqId);
     }
 
@@ -664,7 +668,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
             throw new RuleNotFoundException(rule);
         }
 
-        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(rule, ActionType.Delete));
+        UUID reqId = ConcurrentUtil.get(submitter.submitGlobalRule(tenantContext.tenantId(), rule, ActionType.Delete));
         coordinator.waitForResponse(reqId);
     }
 
@@ -672,7 +676,7 @@ public class KafkaSqlRegistryStorage extends AbstractRegistryStorage implements 
     private void updateArtifactState(ArtifactState currentState, String artifactId, Integer version, ArtifactState newState, EditableArtifactMetaDataDto metaData) {
         ArtifactStateExt.applyState(
             s ->  {
-                UUID reqId = ConcurrentUtil.get(submitter.submitArtifactVersion(artifactId, version, ActionType.Update, newState, metaData));
+                UUID reqId = ConcurrentUtil.get(submitter.submitArtifactVersion(tenantContext.tenantId(), artifactId, version, ActionType.Update, newState, metaData));
                 coordinator.waitForResponse(reqId);
             },
             currentState,
