+++ storage/kafkasql/src/main/java/io/apicurio/registry/storage/impl/kafkasql/keys/MessageKey.java
@@ -37,4 +37,8 @@ public interface MessageKey {
     @JsonIgnore
     public String getPartitionKey();
 
+    /**
+     * Returns the tenantId for this key.
+     */
+    public String getTenantId();
 }
